requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
qdrant-client>=1.4.0
flask>=2.3.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tqdm>=4.65.0
pathlib
concurrent-futures
typing-extensions>=4.5.0
werkzeug>=2.3.0 